Change verification token to global property
Fix the issue where a verification token was assumed to be associated with each workspace instead of with the app itself. The change consists of changing the verification process to check against a global variable instead of the verification token stored with each workspace. The database and associated calls will be changed at a different date.